nvme: Add irqfd, mmap, mmap vs pread comparison
Add additional variations on tests to make use of irqfd for INTx
unmask and mmap access.  Add a simple performance comparison for
mmap vs pread/pwrite MMIO access.  Unfortunately the ctypes
Structure approach to accessing head and tail registers isn't
amenable different strides.  Might need to switch to a slice
approach, but it was slower.  If stride is only ever likely to
be 4 or 8, maybe two structures would work.

With a physical device on bare metal the test will occasionally
get a low number, which is why there's a best-of-3 approach here.
The emulated nvme device in QEMU seems to get stuck pretty
regularly though and requires further debug

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>